webapp_flask/requirements.txt
Flask==2.0.0
Pillow==8.4.0
tensorflow==2.12.0
numpy==1.23.5